implemented _sqrt_recursion

+++ 0x08-recursion/5-sqrt_recursion.c
@@ -5,8 +5,8 @@
  * @n: the number1
  * @m: the number2
  * Return: the square root
- */ 
-int sqrt_fun(int n, int m)
+ */
+long sqrt_fun(long n, long m)
 {
 	if (m >= n / 2)
 		return (-1);
@@ -20,7 +20,7 @@ int sqrt_fun(int n, int m)
  * _sqrt_recursion - calculate the square root of a number
  * @n: the number
  * Return: the square root of a number
- */ 
+ */
 int _sqrt_recursion(int n)
 {
 	return (sqrt_fun(n, 1));
